+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 0.1.1
 * Define `UT` joint scheme
 * Define joint scheme for `Maybe` datatype
+* Define joint scheme for `Either` datatype